Update clustering requests to work for all clients viewing the experiment
