+++ src/redux/actions/cellSets/updateCellSetsClustering.js
@@ -1,57 +1,16 @@
 import {
-  CELL_SETS_ERROR, CELL_SETS_CLUSTERING_UPDATING, CELL_SETS_CLUSTERING_UPDATED,
+  CELL_SETS_ERROR, CELL_SETS_CLUSTERING_UPDATED,
 } from '../../actionTypes/cellSets';
-import { fetchCachedWork } from '../../../utils/cacheRequest';
-import saveCellSets from './saveCellSets';
 
-const REQUEST_TIMEOUT = 30;
-
-const updateCellSetsClustering = (experimentId, resolution) => async (dispatch, getState) => {
-  const {
-    loading, error,
-  } = getState().cellSets;
-
-  const {
-    backendStatus,
-  } = getState().experimentSettings;
-
-  if (loading || error) {
-    return null;
-  }
-
-  const body = {
-    name: 'ClusterCells',
-    cellSetName: 'Louvain clusters',
-    type: 'louvain',
-    cellSetKey: 'louvain',
-    config: {
-      resolution,
-    },
-  };
-
-  dispatch({
-    type: CELL_SETS_CLUSTERING_UPDATING,
-  });
-
+const updateCellSetsClustering = (experimentId, newCellSets) => async (dispatch) => {
   try {
-    const louvainSets = await fetchCachedWork(
-      experimentId, REQUEST_TIMEOUT, body, backendStatus.status,
-    );
-
-    const newCellSets = [
-      louvainSets,
-    ];
-
-    // TODO Factor out up from this
-    // Resolution changes in ui should update cellSets for everyone
-    await dispatch({
+    dispatch({
       type: CELL_SETS_CLUSTERING_UPDATED,
       payload: {
         experimentId,
         data: newCellSets,
       },
     });
-    dispatch(saveCellSets(experimentId));
   } catch (e) {
     dispatch({
       type: CELL_SETS_ERROR,
